dms: Move functions around.
